+++ Geometry.h
@@ -15,8 +15,8 @@ class Geometry {
 public:
 	Geometry() = default;
 	Geometry(const Geometry&) = delete;
-	Geometry(std::vector<T> vertices, std::vector<unsigned short> indices) : vertices(std::move(vertices)), indices(std::move(indices)) {}
+	Geometry(std::vector<T> vertices, std::vector<unsigned int> indices) : vertices(std::move(vertices)), indices(std::move(indices)) {}
 public:
 	std::vector<T> vertices;
-	std::vector<unsigned short> indices;
+	std::vector<unsigned int> indices;
 };
